refactor: split cli.test.ts by command into individual test files
Break the monolithic cli.test.ts into per-command files under
src/__tests__/commands/ for better organization and readability.

- cli.test.ts → CLI-level tests only (help, version, error format)
- commands/list.test.ts, info.test.ts, doc.test.ts, demo.test.ts,
  token.test.ts, semantic.test.ts, changelog.test.ts, migrate.test.ts,
  bug.test.ts, doctor.test.ts, env.test.ts, lint.test.ts, usage.test.ts
