+++ src/__tests__/commands/bug.test.ts
@@ -1,0 +1,69 @@
+import { describe, it, expect } from 'vitest';
+import { run, runCLI } from '../helper.js';
+
+describe('bug', () => {
+  it('should preview a bug report as text', async () => {
+    const out = await run('bug', '--title', 'Test bug');
+    expect(out).toContain('Repository: ant-design/ant-design');
+    expect(out).toContain('Title: Test bug');
+    expect(out).toContain('--- Issue Body ---');
+    expect(out).toContain('### Reproduction link');
+    expect(out).toContain('To submit, re-run with --submit flag.');
+  });
+
+  it('should preview a bug report as JSON', async () => {
+    const out = await run('bug', '--title', 'Test bug', '--format', 'json');
+    const data = JSON.parse(out);
+    expect(data.repo).toBe('ant-design/ant-design');
+    expect(data.title).toBe('Test bug');
+    expect(data.body).toContain('<!-- generated by ant-design-issue-helper');
+    expect(data.url).toContain('https://github.com/ant-design/ant-design/issues/new');
+  });
+
+  it('should include provided fields in bug report', async () => {
+    const out = await run('bug', '--title', 'Test', '--steps', 'Click button', '--expected', 'Works', '--actual', 'Crashes', '--format', 'json');
+    const data = JSON.parse(out);
+    expect(data.body).toContain('Click button');
+    expect(data.body).toContain('Works');
+    expect(data.body).toContain('Crashes');
+  });
+
+  it('should preview bug report as markdown (raw body)', async () => {
+    const out = await run('bug', '--title', 'Test', '--format', 'markdown');
+    expect(out).toContain('<!-- generated by ant-design-issue-helper');
+    expect(out).not.toContain('Repository:');
+  });
+
+  it('should error when --title is missing for bug', async () => {
+    const result = await runCLI('bug', '--format', 'json');
+    expect(result.exitCode).not.toBe(0);
+    const err = JSON.parse(result.stderr);
+    expect(err.code).toBe('TITLE_REQUIRED');
+    expect(err.message).toContain('--title');
+  });
+});
+
+describe('bug-cli', () => {
+  it('should preview a bug-cli report as text', async () => {
+    const out = await run('bug-cli', '--title', 'CLI test bug');
+    expect(out).toContain('Repository: ant-design/ant-design-cli');
+    expect(out).toContain('Title: CLI test bug');
+    expect(out).toContain('### Description');
+  });
+
+  it('should preview a bug-cli report as JSON', async () => {
+    const out = await run('bug-cli', '--title', 'CLI test bug', '--format', 'json');
+    const data = JSON.parse(out);
+    expect(data.repo).toBe('ant-design/ant-design-cli');
+    expect(data.title).toBe('CLI test bug');
+    expect(data.body).toContain('### Description');
+    expect(data.body).toContain('@ant-design/cli');
+    expect(data.url).toContain('https://github.com/ant-design/ant-design-cli/issues/new');
+  });
+
+  it('should include description in bug-cli report', async () => {
+    const out = await run('bug-cli', '--title', 'Test', '--description', 'Info crashes', '--format', 'json');
+    const data = JSON.parse(out);
+    expect(data.body).toContain('Info crashes');
+  });
+});
